# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión Hoteles",
        "description": "Documentación de API para la gestión de hoteles, habitaciones y acomodaciones.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado con éxito"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso"
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
        "/api/habitaciones": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Listar todas las habitaciones",
                "operationId": "4b49cbbd8ecc18afb765c90c4e6b4831",
                "responses": {
                    "200": {
                        "description": "Lista de habitaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Habitacion"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Crear una nueva habitación",
                "operationId": "cc94829c17c067981877ad89e3ed3e84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel_id",
                                    "tipo_habitacion_id",
                                    "acomodacion_id",
                                    "cantidad"
                                ],
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_habitacion_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "acomodacion_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Habitación creada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Habitacion"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Acomodación no válida para el tipo de habitación"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/habitaciones/{id}": {
            "get": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Mostrar una habitación específica",
                "operationId": "0845386ee6507086931fee84e6f6d0fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Habitacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Actualizar una habitación",
                "operationId": "7e850c1c54a27cb97ff3981ea90bf1b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipo_habitacion_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "acomodacion_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Habitación actualizada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Habitacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Habitaciones"
                ],
                "summary": "Eliminar una habitación",
                "operationId": "2849719b65498bf1b95516db06df8f3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la habitación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Habitación eliminada con éxito"
                    },
                    "404": {
                        "description": "Habitación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Habitacion": {
                "required": [
                    "hotel_id",
                    "tipo_habitacion_id",
                    "acomodacion_id",
                    "cantidad"
                ],
                "properties": {
                    "hotel_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tipo_habitacion_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "acomodacion_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "cantidad": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}